Working: refactor BaseHttpStream out of HttpStream & AsyncHttpStream
