+++ airbyte-cdk/python/airbyte_cdk/sources/streams/http/http.py
@@ -19,6 +19,7 @@ from airbyte_cdk.sources.streams.availability_strategy import AvailabilityStrate
 from airbyte_cdk.sources.streams.call_rate import APIBudget, CachedLimiterSession, LimiterSession
 from airbyte_cdk.sources.streams.core import Stream, StreamData
 from airbyte_cdk.sources.streams.http.availability_strategy import HttpAvailabilityStrategy
+from airbyte_cdk.sources.streams.http.http_base import BaseHttpStream
 from airbyte_cdk.sources.utils.types import JsonType
 from airbyte_cdk.utils.constants import ENV_REQUEST_CACHE_PATH
 from requests.auth import AuthBase
@@ -31,7 +32,7 @@ from .rate_limiting import default_backoff_handler, user_defined_backoff_handler
 BODY_REQUEST_METHODS = ("GET", "POST", "PUT", "PATCH")
 
 
-class HttpStream(Stream, ABC):
+class HttpStream(BaseHttpStream, ABC):
     """
     Base abstract class for an Airbyte Stream using the HTTP protocol. Basic building block for users building an Airbyte source for a HTTP API.
     """
@@ -51,22 +52,6 @@ class HttpStream(Stream, ABC):
             self._session.auth = authenticator
         elif authenticator:
             self._authenticator = authenticator
-
-    @property
-    def cache_filename(self) -> str:
-        """
-        Override if needed. Return the name of cache file
-        Note that if the environment variable REQUEST_CACHE_PATH is not set, the cache will be in-memory only.
-        """
-        return f"{self.name}.sqlite"
-
-    @property
-    def use_cache(self) -> bool:
-        """
-        Override if needed. If True, all records will be cached.
-        Note that if the environment variable REQUEST_CACHE_PATH is not set, the cache will be in-memory only.
-        """
-        return False
 
     def request_session(self) -> requests.Session:
         """
@@ -93,54 +78,8 @@ class HttpStream(Stream, ABC):
             self._session.cache.clear()  # type: ignore # cache.clear is not typed
 
     @property
-    @abstractmethod
-    def url_base(self) -> str:
-        """
-        :return: URL base for the  API endpoint e.g: if you wanted to hit https://myapi.com/v1/some_entity then this should return "https://myapi.com/v1/"
-        """
-
-    @property
-    def http_method(self) -> str:
-        """
-        Override if needed. See get_request_data/get_request_json if using POST/PUT/PATCH.
-        """
-        return "GET"
-
-    @property
-    def raise_on_http_errors(self) -> bool:
-        """
-        Override if needed. If set to False, allows opting-out of raising HTTP code exception.
-        """
-        return True
-
-    @property
-    def max_retries(self) -> Union[int, None]:
-        """
-        Override if needed. Specifies maximum amount of retries for backoff policy. Return None for no limit.
-        """
-        return 5
-
-    @property
-    def max_time(self) -> Union[int, None]:
-        """
-        Override if needed. Specifies maximum total waiting time (in seconds) for backoff policy. Return None for no limit.
-        """
-        return 60 * 10
-
-    @property
-    def retry_factor(self) -> float:
-        """
-        Override if needed. Specifies factor for backoff policy.
-        """
-        return 5
-
-    @property
     def authenticator(self) -> HttpAuthenticator:
         return self._authenticator
-
-    @property
-    def availability_strategy(self) -> Optional[AvailabilityStrategy]:
-        return HttpAvailabilityStrategy()
 
     @abstractmethod
     def next_page_token(self, response: requests.Response) -> Optional[Mapping[str, Any]]:
@@ -163,73 +102,6 @@ class HttpStream(Stream, ABC):
         """
         Returns the URL path for the API endpoint e.g: if you wanted to hit https://myapi.com/v1/some_entity then this should return "some_entity"
         """
-
-    def request_params(
-        self,
-        stream_state: Optional[Mapping[str, Any]],
-        stream_slice: Optional[Mapping[str, Any]] = None,
-        next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> MutableMapping[str, Any]:
-        """
-        Override this method to define the query parameters that should be set on an outgoing HTTP request given the inputs.
-
-        E.g: you might want to define query parameters for paging if next_page_token is not None.
-        """
-        return {}
-
-    def request_headers(
-        self,
-        stream_state: Optional[Mapping[str, Any]],
-        stream_slice: Optional[Mapping[str, Any]] = None,
-        next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> Mapping[str, Any]:
-        """
-        Override to return any non-auth headers. Authentication headers will overwrite any overlapping headers returned from this method.
-        """
-        return {}
-
-    def request_body_data(
-        self,
-        stream_state: Optional[Mapping[str, Any]],
-        stream_slice: Optional[Mapping[str, Any]] = None,
-        next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> Optional[Union[Mapping[str, Any], str]]:
-        """
-        Override when creating POST/PUT/PATCH requests to populate the body of the request with a non-JSON payload.
-
-        If returns a ready text that it will be sent as is.
-        If returns a dict that it will be converted to a urlencoded form.
-        E.g. {"key1": "value1", "key2": "value2"} => "key1=value1&key2=value2"
-
-        At the same time only one of the 'request_body_data' and 'request_body_json' functions can be overridden.
-        """
-        return None
-
-    def request_body_json(
-        self,
-        stream_state: Optional[Mapping[str, Any]],
-        stream_slice: Optional[Mapping[str, Any]] = None,
-        next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> Optional[Mapping[str, Any]]:
-        """
-        Override when creating POST/PUT/PATCH requests to populate the body of the request with a JSON payload.
-
-        At the same time only one of the 'request_body_data' and 'request_body_json' functions can be overridden.
-        """
-        return None
-
-    def request_kwargs(
-        self,
-        stream_state: Optional[Mapping[str, Any]],
-        stream_slice: Optional[Mapping[str, Any]] = None,
-        next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> Mapping[str, Any]:
-        """
-        Override to return a mapping of keyword arguments to be used when creating the HTTP request.
-        Any option listed in https://docs.python-requests.org/en/latest/api/#requests.adapters.BaseAdapter.send for can be returned from
-        this method. Note that these options do not conflict with request-level options such as headers, request params, etc..
-        """
-        return {}
 
     @abstractmethod
     def parse_response(
@@ -284,24 +156,6 @@ class HttpStream(Stream, ABC):
         """
         return ""
 
-    def must_deduplicate_query_params(self) -> bool:
-        return False
-
-    def deduplicate_query_params(self, url: str, params: Optional[Mapping[str, Any]]) -> Mapping[str, Any]:
-        """
-        Remove query parameters from params mapping if they are already encoded in the URL.
-        :param url: URL with
-        :param params:
-        :return:
-        """
-        if params is None:
-            params = {}
-        query_string = urllib.parse.urlparse(url).query
-        query_dict = {k: v[0] for k, v in urllib.parse.parse_qs(query_string).items()}
-
-        duplicate_keys_with_same_value = {k for k in query_dict.keys() if str(params.get(k)) == str(query_dict[k])}
-        return {k: v for k, v in params.items() if k not in duplicate_keys_with_same_value}
-
     def _create_prepared_request(
         self,
         path: str,
@@ -328,10 +182,6 @@ class HttpStream(Stream, ABC):
         prepared_request: requests.PreparedRequest = self._session.prepare_request(requests.Request(**args))
 
         return prepared_request
-
-    @classmethod
-    def _join_url(cls, url_base: str, path: str) -> str:
-        return urljoin(url_base, path)
 
     def _send(self, request: requests.PreparedRequest, request_kwargs: Mapping[str, Any]) -> requests.Response:
         """
