+++ airbyte-cdk/python/airbyte_cdk/sources/streams/http/http_async.py
@@ -18,7 +18,7 @@ from airbyte_cdk.sources.streams.async_call_rate import AsyncCachedLimiterSessio
 from airbyte_cdk.sources.streams.call_rate import APIBudget
 from airbyte_cdk.sources.streams.core import StreamData
 from airbyte_cdk.sources.streams.core_async import AsyncStream
-from airbyte_cdk.sources.streams.http import HttpStream
+from airbyte_cdk.sources.streams.http.http_base import BaseHttpStream
 from airbyte_cdk.sources.streams.http.exceptions_async import DefaultBackoffException, RequestBodyException, UserDefinedBackoffException
 from airbyte_cdk.sources.utils.types import JsonType
 from airbyte_cdk.utils.constants import ENV_REQUEST_CACHE_PATH
@@ -31,7 +31,7 @@ BODY_REQUEST_METHODS = ("GET", "POST", "PUT", "PATCH")
 T = TypeVar("T")
 
 
-class AsyncHttpStream(HttpStream, AsyncStream, ABC):
+class AsyncHttpStream(BaseHttpStream, AsyncStream, ABC):
     """
     Base abstract class for an Airbyte Stream using the HTTP protocol with asyncio.
